+++ 02_ft_printf/bonus/ft_print_uxXp_bonus.c
@@ -6,13 +6,13 @@
 /*   By: jgwon <jgwon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/29 19:44:03 by jgwon             #+#    #+#             */
-/*   Updated: 2022/08/29 21:12:22 by jgwon            ###   ########.fr       */
+/*   Updated: 2022/08/30 19:45:48 by jgwon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf_bonus.h"
 
-int	ft_print_sharp(unsigned long long d, t_info *info)
+int	ft_print_sharp(unsigned long d, t_info *info)
 {
 	int	cnt;
 
@@ -38,7 +38,7 @@ int	ft_print_sharp(unsigned long long d, t_info *info)
 	return (cnt);
 }
 
-int	ft_print_uxp(unsigned long long d, t_info *info)
+int	ft_print_uxp(unsigned long d, t_info *info)
 {
 	int	cnt;
